finish registration + outsource address in a new file
