+++ src/main/webapp/home.xhtml
@@ -5,6 +5,7 @@
     <title>Success</title>
 </h:head>
 <h:body>
-    <h:outputText value="Login successful! Welcome, #{authentifizierung.username}." />
+    <h:outputText value="Welcome, #{authentifizierung.username}. (Hier kann der Home-Bildschirm angezeigt werden)" />
 </h:body>
 </html>
+<!-- TODO: Implement manager so we can access the current user across the application>
